except/requirements.txt
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-modelcluster==4.3
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.2
draftjs-exporter==2.1.5
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
mysqlclient==1.4.2.post1
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
six==1.12.0
static3==0.7.0
Unidecode==1.0.23
urllib3==1.24.1
wagtail==2.5.1
webencodings==0.5.1
Willow==1.1
wagtailmenus==2.13.1
wagtail-modeltranslation==0.10.2
preview-generator==0.10